Try out emitting defered events after the new event.
The obvious benefit is that if you are waiting for something else to
change in the model, that thing can occur before the thing that you said
you want to defer. (eg, if you are waiting for a relation to be joined
before you handle config-changed, then don't check config-changed
immediately before relation-joined.)
The odd downside is that any defered event is immediately emitted a
second time, before the process exits.