Some exploratory work for creating symlinks early.
We could do it, but it changes how we know the name of a hook. Right now
it is by defining

class CharmEvents(ObjectEvents):
  name = EventSource(HookEventClass)

That name is the attribute of the ObjectEvents class. The HookEventClass
itself doesn't know its hook name. We can change that, but it is enough
different that we should confirm that we really want that first.